Refactored losses.base into separate files
