+++ distances.py
@@ -2,7 +2,6 @@
 # -*- coding: utf-8 -*-
 import torch
 import torch.nn.functional as F
-import numpy as np
 
 
 class Distance:
