+++ losses/triplet.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import numpy as np
